Move CI_TOKEN and related code from AbstractVersionControlService to GitLabService
